requirements.txt
tensorflow
flask<=2.3
flask-wtf
flask_sqlalchemy
flask_login
flask-migrate
pydicom
matplotlib
numpy
tqdm
dicom2nifti
nibabel
Pillow
scipy
ipykernel
ipywidgets
opencv-python
jupyter
notebook
antspyx
SimpleITK
antspynet
numpy-stl
vedo
patchify
pyrebase
customtkinter
sv_ttk
CustomTkinterTitlebar
reportlab
plotly
trame
trame-vuetify trame-vtk
vtk
